Fix `Precision` in `Decimal` values
 + Revised the precision of the decimal values to show only two values after the decimal point.
 + It will only convert the expressions to "BigInt" only if it's not a decimal.

Earlier, there was a problem when calculating expressions due to floating point.
As the floating point can't be converted to "BigInt", it throws error when involved with floating points.
Now, if there was a floating point, the program will parse it as a "Number".
This change made the problem to disappear.