Fix bug in parantheses deletion & inclusion
 + Added a logic, to check the size of the "operandOne" while deletion.
 + If it's empty, then parantheses will be removed.
 + It the result is positive, then no parantheses would be included.
 + If the value is negative, and clicked the "+/-" it adds parantheses alone.

Earlier, if there was only "(-)" and if deletion is carried out, it will leave out "()".
Now, a logic is in place to check the length and if there are parantheses, it will be removed.

Earlier, if there was already a negative value from result, and if "+/-" clicked, it will add another minus "(--2)".
Now, a logic is in place to check if there was an existing minus.

Earlier, if there was parantheses while calculation, it will be forwarded to the result.
Now, it will be removed if the result is positive.